requirements.txt
pydub 
numpy
Pillow
google-api-python-client
google_auth_oauthlib
